initial vae changes
