+++ capsa/bias/__init__.py
@@ -1,1 +1,2 @@
-from .histogram import HistogramWrapper, HistogramCallback
+from .histogram import HistogramWrapper
+from .histogram_callback import HistogramCallback
